Lot of work on onboarding, still not finished

+++ src/router.js
@@ -19,7 +19,6 @@ import Dashboard from './components/Dashboard'
 import Session from './components/Session'
 import ListSessions from './components/ListSessions'
 import Action from './components/Action'
-import Profile from './components/Profile'
 import Schedule from './components/Schedule'
 import Resources from './components/Resources'
 import Feedback from './components/Feedback'
@@ -50,7 +49,7 @@ const routes = [
   { path: '/feedback', component: Feedback, meta: { protected: true } },
   { path: '/action/:action/:data?', component: Action, meta: { bypassOnboarding: true } },
   { path: '/onboarding/:step?', component: Onboarding, meta: { protected: true } },
-  { path: '/profile', component: Profile, meta: { protected: true, bypassOnboarding: true } }
+  { path: '/profile', redirect: '/onboarding/profile' }
 ]
 
 const router = new VueRouter({
@@ -95,7 +94,7 @@ router.beforeEach((to, from, next) => {
 // If endpoint returns 401, redirect to login (except for requests to get user's session)
 Vue.http.interceptors.push(function(request, next) {
   next(function(response){
-    if (response.status === 401 && request.url.indexOf('/api/user') === -1){
+    if (response.status === 401 && !(request.url.indexOf('/api/user') !== -1 && request.method === 'GET')){
       AuthService.removeUser();
       router.push('/login?401=true');
     }
